backend/requirements.txt
flask
pandas
gunicorn
bs4
requests
html5lib
python-dotenv
pymongo
pyjwt